[data] allow ConcurrencyCapBackpressurePolicy._cap_multiplier to be set to 1.0  (#41222)
Allow ConcurrencyCapBackpressurePolicy._cap_multiplier to be set to 1.0. This means that the concurrency will stay the same. No ramping up. This is because for some workloads, we need an upper bound of concurrency cap to make it stable. Before introducing new configs for control concurrency. Add this as a temporary workaround.
---------

Signed-off-by: Hao Chen <chenh1024@gmail.com>