[Serve] Scale `initial_replicas` by `target_capacity` during rollouts (#41201)
Users can set initial_replicas in their autoscaling_config to specify the number of replicas that a deployment should start with. This change defines the behavior for initial_replicas during rollouts where target_capacity is specified:

If target_capacity is increasing for a config, initial_replicas is scaled by target_capacity, and this value is used as a floor for the number of replicas.
If target_capacity is decreasing for a config, initial_replicas is ignored, and min_replicas is treated as a floor.
Additional Changes:

Currently, the Serve autoscaler calculates the number of replicas by running the autoscaling algorithm between min_replicas and max_replicas and then scaling that decision by the target_capacity. This PR makes the autoscaler clip that number by scaled min_replicas/max_replicas bounds instead of scaling the number directly.
Motivation: Suppose target_capacity is 50. Semantically, that means Serve should be handling 50% of traffic as usual. If that's the case, then the autoscaler is deciding num_replicas based on 50% of the traffic already. There's no reason to scale that decision down by another 50%. Instead, it should be clipped by the scaled bounds.