[CI] [Cluster] Fix example GCP GPU/docker example cluster YAML file (#41134)
Fixes the broken GCP GPU docker example YAML by updating the image and cuda version. (Thanks @PaulFenton for the fix!)

Adds a release test that checks that ray up works with this YAML file (it didn't work before).

Switched the image used in the YAML to ray instead of ray-ml because ray-ml gave errors with running out of disk space, and because it makes the test run faster. If it's important to make ray-ml the default, we can probably just increase diskSizeGb and test it again.

Related issue number
Closes #35266

---------

Signed-off-by: Archit Kulkarni <archit@anyscale.com>