[Core] Optimize next/anext performance for streaming generator (#41270)
Currently, when the worker has a lot of yield and next in a generator, it is bottlenecked by the performance of yield and next itself.

This PR optimizes the performance if next. next was slow because we always call ray.wait, which was pretty expensive.

wait is not necessary when an object is actually already ready. If it is not ready yet, the overhead of wait is trivial. This PR fixes the performance overhead by not calling ray.wait if the object is already ready when we peak it (we can figure it out because we know when the object is reported, meaning it is ready).

After this PR, the throughput of tests are doubled.

After this yield becomes the bottleneck. Unfortunately, yield is hard to optimize because it is due to serialization. We will deal with it in the longer term.