[Core] Make instance_of_cause work with ray internal error (#41294)
Imagine you have a controller actor with a worker actor and have a call to controller actor (i.e., Ray tune).

Controller -> Worker

ray.get(controller.get_result.remote())
If Worker fails, ray.get will raise an exception RayTaskError.

Normally when Ray raises an exception, the exception is a subclass of both RayTaskError and the original exception (returned from RayTaskError.as_instance_of_cause). However, when the Ray's internal exception is raised, it doesn't return the original internal error, but only RayTaskError.

This behavior is confusing, and for users who'd like to obtain the correct "root error", they need to manually access internal private attribute RayTaskError.cause.

The motivation of the change above ^ was because of the following two issues.
#5862
#5870
TL;DR some internal error contains an object id, and we couldn't obtain it. In the master, this is not true anymore, and there's no reason to do this anymore.

After this PR, all the internal error will also be contained from an exception, and user can easily access the raised exception or use exception.as_instance_of_cause() to access the original exception properly.