[cherry-pick][docs] Fix algolia autocomplete overlap (#41200) (#41284)
* [Doc] Fix algolia autocomplete overlap (#41200)

This PR fixes an issue where the algolia searchbox autocomplete menu would be "overlapped" by the main content area of the page.

In actuality this was not a z-index issue, but an issue with the overflow style on the #site-navigation bar. A rule setting a fixed size for the autocomplete box was also removed, along with a rule forcing the autocomplete box to always be scrollable - neither of these is needed here.

Signed-off-by: pdmurray <peynmurray@gmail.com>

* add new line to kick of CI

Signed-off-by: angelinalg <122562471+angelinalg@users.noreply.github.com>

---------

Signed-off-by: pdmurray <peynmurray@gmail.com>
Signed-off-by: angelinalg <122562471+angelinalg@users.noreply.github.com>
Co-authored-by: Peyton Murray <peynmurray@gmail.com>