[Serve] Support automatic conversion for Pydantic 2 models in the args builder (#41542)
Users can annotate their args builder with a Pydantic model to automatically convert config args to a Pydantic model. However, if they use a Pydantic v2 model, the config args will not be converted to a model. The args remain a dict.

This change makes the args builder automatically convert the dict to a model, no matter what Pydantic version the model was defined in.

---------

Signed-off-by: Shreyas Krishnaswamy <shrekris@anyscale.com>