[core] Add GPU support for MPI (#41349)
By default, each mpi worker will see all GPUs on the node. This will cause issue when the num_gpus is given for tasks or actors, because their view of the GPU visible is different.

This PR fixed it. It requires the user to call mpi_init inside the function before any operations.