[data] add task filter for node killer (#41099)
Adds a node filter to NodeKillerActor to target nodes with specific tasks running on them. Used for targeting data+train nodes for training benchmarks.

Example usage:

>>> python setup_chaos.py --task-names "ReadRange->MapBatches(<lambda>)"
...
>>> ray.data.range(...).map_batches(lambda x: x).materialize()

When running with release tests, we should try to keep the test short because this uses the state api to query running tasks, which have a limit of 10k tasks (#33547). So currently we only run this with our train benchmarks with --num-epochs 1

---------

Signed-off-by: Andrew Xue <andewzxue@gmail.com>
Signed-off-by: can <can@anyscale.com>