[Ray Core] Make worker_process_setup_hook work with strings instead of Python functions #41096 (#41479)
This PR makes worker_process_setup_hook with work with string.

The implementation is as follow;

We always append a special identifier if funcs are imported. If funcs are imported, we import it from the GCS and execute.
If string is passed, we just pass the function without the special identifier. In this case, we load the module and run the setup function.