[Core][2/3] Make streaming generator public API #38784 (#41436)
This is the second PR of 3 steps.

All sort of implementation work [done]
Rename StreamingObjectRefGenerator and annotate it as a public API (and move it to init.py) <---- this PR
Typing support [future]

This PR renames StreamingObjectRefGenerator - > ObjectRefGenerator. This also means the existing dynamic generator's (num_returns="dynamic") return type will become ObjectRefGenerator -> DynamicObjectRefGenerator, which is a breaking change, but we will anyway deprecate this API. I am open for a different name if we'd like to keep the perfect backward compatibility.