[Ray Debugger] add debugpy as the ray debugger (#328)
## Context
We are integrating ray debugger with workspace web-vscode. There are
three major components playing different roles in the workflow:
1. debugger UI (vscode extension): it's the UI to show status and to
connect vscode debugger with ray debugger.
2. ray breakpoint handler (rpdb): it pauses a worker task when
breakpoint is hit or unhandled exception is captured, and waits debugger
client to connect. Also, it updates the state of the tasks to Ray GCS.
3. Ray GCS: Ray Core API to keep track of current state of the tasks

This PR is covering the `ray debugger` part.

## Changes
This PR is refactoring the rpdb part of Sofian's prototyping PR:
https://github.com/anyscale/runtime/pull/317

1. enabled debugpy as the ray debugger for breakpoint and post_mortem
debugging
2. added flag `RAY_DEBUG=1` to enable debugpy. If `RAY_DEBUG` is not set
and `RAY_PDB` is set, then rpdb will be used.
3. used state api to save worker debugging port.

## Test plan

I manually tested that the workers can be paused at breakpoints and
unhandled exceptions (post_mortem)

---------

Signed-off-by: bhuang <bhuang@anyscale.com>
Co-authored-by: adf <adf@adf.adsf>