[Data] Partial fix for Dataset.context not being sealed after creation (#41569)
`Dataset.context` should be sealed the first time the Dataset is created. But if a new operator is applied to the dataset, the new global DataContext will be saved again to the Dataset. 

This bug prevents using different DataContexts for training and validation datasets in a training job.

Note this PR only fixes the issue when multiple datasets are created in the process but will be running in different processes. If they run in the same process, it's still a bug, see https://github.com/ray-project/ray/issues/41573.

---------

Signed-off-by: Hao Chen <chenh1024@gmail.com>