[docs][data][core][tune] remove mention of Many Model Training (#41682)
After much discussion, we are removing the concept and examples associated with MMT. We had three different ways of running it (Core, Data, Tune). We no longer recommend Tune. Although this is a valid workload, the upkeep of the examples and docs outweigh the value, based on page views. When there is more traction on this type of use case, we can create a more up-to-date example.
@pcmoritz @richardliaw @matthewdeng

Signed-off-by: angelinalg <122562471+angelinalg@users.noreply.github.com>