[2.9.0 Release] [Doc] Update Dask-on-Ray dependencies for 2.9.0 and 2.8.0 (#41639)
Updates the Dask-on-ray compatibility table based on the following instruction:

Get the latest compatible Dask version from here and create a PR to master to update the doc. The PR needs to be cherry picked to the release branch as well. The file is moved to doc/source/ray-more-libs/dask-on-ray.rst

The 2.8.0 update was missing; this PR includes it as well.

These Ray versions do not support Python 3.7, so we don't have an entry for Python 3.7 like we do for older Ray versions in the table.

---------

Signed-off-by: Archit Kulkarni <archit@anyscale.com>