[Serve] [Core] Container runtime env support (#40419) (#41716)
Cherry pick of #40419.

This is a feature we want to get out for 2.9, quite a few OSS users are interested in trying it out (see thread). We want to reach out to users to try it out and gather feedback once it’s released in 2.9 (otherwise we will have to wait for 2.10). It’s also pretty low risk, the changes are confined to the container part of runtime_env, and most of the PR is actually adding tests.

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>