[serve] Controller recovery edge case (#40874)
When the controller recovers after crashing, it collects a list of replica actor names and recovers them one by one. But while it's recovering the replicas, the replicas could have died for some reason (e.g. killed), so that by the time the controller gets around to a certain replica, it's already dead. But the controller assumes the actor must be alive because it collected the replica names from live actors. This PR addresses this issue: if the controller fails to get a handle to the actor during recovery, simply mark that actor as dead (basically "skip" recovering that replica).

---------

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>