[Doc] Make linkcheck less aggressive when run on PR branches (#41813)
This PR makes the linkcheck CI job only check the ray-project github URLs, since those are the ones we have control over. If these don't pass, chances are you can't even push. This should increase the success rate of the CI job, while still performing the most important checks that we need on the docs.

Signed-off-by: pdmurray <peynmurray@gmail.com>