[Core] Add pre-commit configuration (#41500)
This PR adds a pre-commit configuration to Ray. This has a number of advantages over the current pre-commit hooks over format.sh:

More developers are familiar with pre-commit from other projects, decreasing maintenance burden. No special knowledge of Ray dev work flow required to maintain this.
Isolated environments for these pre-commit hooks mean that there is not need to check what version of a tool the developer has (or doesn't have) on their machine; pre-commit just gets the right version and uses it in an isolated environment without any checks, warnings, or version conflicts; e.g.
image

If we fully migrate to pre-commit, we won't need lint-requirements.txt because the lint requirements will be fully specified in .pre-commit-config.yml. This should simplify our CI system and our build constraints.
Users wouldn't have to download the google-java-formatter as much, as it will be cached by the hook rather than placed in /tmp/
It's easier to upgrade hooks with pre-commit autoupdate, or by updating the yaml config manually
It's easier to add new hooks: just edit the yaml config to make a new hook, instead of learning how the current lint system works with ci/lint/format.sh. This would be useful because there's a lot of lint checks that are run when a PR is made but not locally - see the ci/lint/ directory for examples of this.
Note that this config is not enforced, but in the future we may remove the current lint system in favor of pre-commit.

I'm open to turning on more hooks, or turning some off if people prefer; note also that this PR doesn't include changes suggested by this config. For the most part, the changes that get applied when you run pre-commit with this config due to trim-trailing-whitespace and end-of-file-fixer. There are a few other issues which are legitimate concerns that I'll make issues for later.

Changes
Added a .pre-commit-config.yaml that matches format.sh as closely as possible
Broke check_docstyle out into its own script, which is run as a local hook
Added a .prettierrc.toml; currently prettier only runs on files in doc/

Signed-off-by: pdmurray <peynmurray@gmail.com>