[Core] Disallow callable for job submission api (#41788) (#41947)
Passing callable type to job submission API should be an anti pattern because you need to serialize the callable from the job submission side, and the cluster side can easily have a different environment when you deserialize it.

The best practice is we only allow string (module name) for worker_process_setup_hook when it is passed from job submission code. Note that worker_process_setup_hook is allowed to be overwritten (without affecting other runtime env) if it is specified via ray.init.