Update d4rl_reader.py
for d4rl, we need the gym, not the gymnasium.

if we use import gymnasium as gym, we will get the exception: gymnasium.error.NameNotFound: Environment `maze2d-umaze` doesn't exist. 

this environment is the classic env in d4rl.

if we change it to import gym, then it will be ok.

Signed-off-by: retroactively <41630114+retroactively@users.noreply.github.com>