[Cluster launcher] [vSphere] Refactor the vSphere sdk provider (#41911)
Similar with our previous change, in this change we continue to wrap the operations we conduct against the vSphere cluster. We seal the operations in vsphere_sdk_provider.py. We only keep the main logic in node_provider.py

Because of this, some UT for node provider need a change, so we split the original UT into 2 files, to cover the code.

---------

Signed-off-by: Chen Jing <jingch@vmware.com>