[serve] Run `UserCallableWrapper` code on separate event loop (thread) (#42298)
Run all user code in the UserCallableWrapper on a separate event loop to avoid blocking the main replica loop.

After this PR is merged, I will be unblocked to productionize the prototype for late binding / single-RPC routing. This change is a requirement as we need to ensure the call to get the queue length/start executing a request (1) cannot hang due to user code blocking the event loop and (2) run on the same thread as each other.

This has a more general benefit in that it avoids having user code able to fully block our code in the replica (can help improve observability quite a lot) and makes the concurrency model more explicit. Currently some of the methods implicitly run on separate threads/event loops (using concurrency_group support) which makes it very easy to have subtle race conditions. For example user health check methods are not currently thread safe because they're actually called on a separate event loop.

I've removed the usage of concurrency_groups in this PR because they are no longer needed. This actually includes a minor bugfix for custom user check_health methods that can be considered a behavior change. Previously, the check_health method was run on a separate asyncio loop/thread as that method was in another concurrency group. Now, it runs on the same thread as the rest of the user code. This means users must avoid blocking the asyncio loop or else their custom health checks may time out. To avoid possibly causing problems for users without a custom health check, I've elected to skip interacting with the user code thread entirely in that case.

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>
Co-authored-by: shrekris-anyscale <92341594+shrekris-anyscale@users.noreply.github.com>
Co-authored-by: Cindy Zhang <cindyzyx9@gmail.com>