[serve] Maintain replica queue length directly instead of using `_get_actor_call_stats` (#42485)
Removes the reliance on the sketch _get_actor_call_stats API (and removes that API entirely as it was not used elsewhere). Instead, we maintain the count of ongoing requests directly as a counter in the ReplicaMetricsManager.

This is possible after #42298 because all user code now runs on a separate thread so the main actor asyncio loop will be blocked by synchronous user code (unless it holds the GIL, which was not handled previously anyways).

Due to this change, we can no longer maintain and report the serve_replica_pending_queries metric. This metric was almost always 0 and effectively useless, so I have elected to completely remove it after offline discussion with the team.

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>
Co-authored-by: shrekris-anyscale <92341594+shrekris-anyscale@users.noreply.github.com>
Co-authored-by: Cindy Zhang <cindyzyx9@gmail.com>