[Data] Log slow UDF warning to `ray-data.log` file only (#42614)
Currently, Ray Data emits a warning when a UDF takes longer than 10 seconds, which is sent to both stdout and the Ray Data specific log file. We reduce the verbosity of the warning by only logging to the file, since this is not necessarily an indication of something wrong (e.g. intensive tasks will naturally take longer).

Signed-off-by: Scott Lee <sjl@anyscale.com>
Signed-off-by: khluu <khluu000@gmail.com>