shorten bazel output directories even more on windows (#42566)
The AUTOMATED_BUILD environment variable is set on the conda-forge builds to solve various build problems when building there. One problem that came up in the 2.9.0 update is that the windows path is too long, which resulted in hard-to-debug build problems. This effectively does the fix for #42098 (setting up .bazelrc with startup --output_user_root=d:/tmp) programatically without involving .bazelrc.

The code path will not be hit in the buildkite CI tests.

---------

Signed-off-by: Matti Picus <matti.picus@gmail.com>
Signed-off-by: mattip <matti.picus@gmail.com>

Signed-off-by: wesley <rademakerwesley@gmail.com>