[train+tune] Make path joining OS-agnostic by using `Path.as_posix` over `os.path.join` (#42037)
`os.path.join` uses an OS-dependent separator (`\` on windows and `/` on posix systems), which causes some issues when `os.path.join` is used in conjunction with `Path`. The combination can result in a mix of `/` and `\` when running Ray Train/Tune on windows. This runs into issues when passing these paths into `pyarrow.fs`, leading to issues such as `FileNotFound`.

---------

Signed-off-by: n3011 <mrinal.haloi11@gmail.com>
Signed-off-by: Ishant Mrinal <33053278+n30111@users.noreply.github.com>
Signed-off-by: Justin Yu <justinvyu@anyscale.com>
Co-authored-by: n3011 <mrinal.haloi11@gmail.com>
Co-authored-by: Justin Yu <justinvyu@anyscale.com>