[serve] stabilize test_standalone_3
When one test fails in `test_standalone_3.py`, all the tests after it fail because `serve.shutdown()` doesn't get executed at the end of the test, and so the handle the client has to the controller isn't reset. Then in the next test, the client tries to ping the controller at the old handle and fails.

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>