[Data] Support pandas 2.2.0 for pandas tensor extensions (#42892)
`ExtensionArrayFormatter` is replaced with `_ ExtensionArrayFormatter` in pandas 2.2.0. If Ray Data users have this version of pandas installed, it will result in the error in https://github.com/ray-project/ray/issues/42842.

This PR adds the case to support the change to `ExtensionArrayFormatter`. Tested locally with pandas 2.2.0 and Ray Data usage works without the error. 

Signed-off-by: Scott Lee <sjl@anyscale.com>