[core][client] Remove protocol version, instead just compare Ray and Python versions. (#42760)
In a regular Ray worker init, (python version, ray version) is compared against the cluster. In Ray Client, however, a (python version, protocol_version) is compared. This makes it a looser requirement for the client connection than for the workers.

This commit removes that and unifies version check between both sides to (python version, ray version). This helps avoiding any confusion in version mismatch.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>