Fix num blocks local, remote, and unknown location are all 0 (#42888)
When printing the stats for a dataset, the counts of blocks that are local, remote or unknown are sometimes all 0.

This occurs when ctx.enable_get_object_locations_for_metrics is false on this line (which is the default behavior). This PR hides these stats if ctx.enable_get_object_locations_for_metrics is false as they will all be 0. It also refactors the _calculate_ref_hits method to make it more clear that if set to false the return will be all zero counts.

Closes #42796, closes #31144

---------

Signed-off-by: Matthew Owen <mowen@anyscale.com>
Signed-off-by: Matthew Owen <omatthew98@berkeley.edu>
Co-authored-by: Scott Lee <scottjlee@users.noreply.github.com>