[Data] Avoid `TypeError` when destructing `Dataset` object (#43014)
You might see a `TypeError` when your Ray Data program exits:

```
Exception ignored in: <function Dataset.__del__ at 0x2aad7d3a0>
Traceback (most recent call last):
  File "/Users/arun/Applications/miniconda3/envs/dc/lib/python3.9/site-packages/ray/data/dataset.py", line 5148, in __del__
TypeError: 'NoneType' object is not callable
```

The reason for this is that we don't correctly check if the `ray` module exists. Currently, we check if `ray is not None`:

https://github.com/ray-project/ray/blob/ba3cfa0eb0daaf9ecdb17edc1c4fa4bbeb4d9dd8/python/ray/data/dataset.py#L5062

But `ray` might evaluate to `<module None from None>`, which is not `None`. So, we evaluate `ray.is_initialized()` and get a `TypeError`. This PR fixes the issue by adding a more robust check.

---------

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>