[Data] Time streaming exec scheduling (#43112)
Currently we are not timing how much time is spent during scheduling in the streaming executor. This times the total `process_time` for the scheduling steps / calls to `_scheduling_loop_step`. This stat is included in DatasetStats and a later PR will include this and other StreamingExecutor stats into the DatasetStatsSummary.

Signed-off-by: Matthew Owen <mowen@anyscale.com>