[object_buffer_pool] reduce mutex lock scope in WriteChunk (#43150)
Object store network transfer performance is slow, and we observe a periodic burst followed by a gap in the network usage.

A burst of inbound network traffic occurs at the beginning of each ray.get(obj_refs) call, then there is a wide-gap of unused network traffic and then a subsequent burst of network traffic in the next ray.get(obj_refs) call.

This looks like a processing bottleneck when payloads are received over the network on the Pull side.

We dug into the code in object_manager and found the following:

Objects are transferred in chunks of size: 5 MiB
When a PushRequest is received for a chunk, it is processed by ObjectManager::HandlePush
Which internally calls the ObjectManager::ReceiveObjectChunk. This results in a call to the function ObjectBufferPool::WriteChunk.
The WriteChunk function is mutex guarded throughout its execution.
This includes the std::memcpy call for a 5MiB payload
This pool_mutex_ lock is shared by all object_ids being received over the network
Which makes us believe that even if chunks for different ObjectId are received in parallel over the network they are written sequentially. Which would explain why we see a burst in the network usage followed by a hole in the network usage

Changes
Transition the chunk from REFERENCED to SEALED before releasing the lock
Perform an unguarded memcpy of the chunk into the buffer
Reacquire the mutex lock and perform object_id level Seal and Release decisions